services/log-parser/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
pyyaml
jsonpath-ng
